requirements.txt
openai
python-dotenv
python-telegram-bot
psycopg2-binary
